split tests a lot of fixes/todos
